Improve class inheritance for types and sessions
Ref. #208
